NMR-4926: move all vendor specific classes to subpackages
